+++ isce2_topsapp/topsapp_params.py
@@ -1,0 +1,31 @@
+from pydantic import BaseModel
+from typing import List
+
+
+class topsappParams(BaseModel):
+    reference_scenes: List[str]
+    secondary_scenes: List[str]
+    frame_id: int
+    estimate_ionosphere_delay: bool = True
+    compute_solid_earth_tide: bool = True
+    output_resolution: int = 90
+    unfiltered_coherence: bool = True
+    goldstein_filter_power: float = .4
+    dense_offsets: bool = False
+    wrapped_phase_layer: bool = False
+    esd_coherence_threshold: float = -1
+
+    def is_standard_gunw_product(self) -> bool:
+        """Version 3+"""
+        checks = [
+            self.estimate_ionosphere_delay == True,
+            self.frame_id != -1,
+            self.compute_solid_earth_tide == True,
+            self.dense_offsets == True,
+            self.unfiltered_coherence == True,
+            self.esd_coherence_threshold == -1,
+            self.wrapped_phase_layer == False,
+            self.goldstein_filter_power == .4,
+            self.output_resolution == 90,
+        ]
+        return all(checks)